Some clean-up
