+++ sparse-core/src/distributed/indexable_vector.rs
@@ -2,10 +2,10 @@
 use mpi::traits::*;
 use num::{Float, Zero};
 use sparse_traits::types::{Error, Result};
-use sparse_traits::IndexableVector;
+use sparse_traits::linalg::IndexableVector;
 use sparse_traits::Scalar;
 use sparse_traits::{IndexLayout, IndexType};
-use sparse_traits::{Inner, Norm1, Norm2, NormInf, SquareSum};
+use sparse_traits::linalg::{Inner, Norm1, Norm2, NormInf, SquareSum};
 
 use super::index_layout::DistributedIndexLayout;
 
@@ -172,82 +172,3 @@ where
         global_result
     }
 }
-
-
-/* impl<T: Scalar> IndexableVector for LocalIndexableVector<'_, T> {
-    type Ind = LocalIndexLayout;
-    type Iter<'b> = std::slice::Iter<'b, T> where Self: 'b;
-
-    type IterMut<'b> = std::slice::IterMut<'b, T> where Self: 'b;
-    type T = T;
-
-    fn get(&self, index: IndexType) -> Option<&Self::T> {
-        self.data.get(index)
-    }
-
-    fn get_mut(&mut self, index: IndexType) -> Option<&mut Self::T> {
-        self.data.get_mut(index)
-    }
-
-    unsafe fn get_unchecked(&self, index: IndexType) -> &Self::T {
-        self.data.get_unchecked(index)
-    }
-
-    unsafe fn get_unchecked_mut(&mut self, index: IndexType) -> &mut Self::T {
-        self.data.get_unchecked_mut(index)
-    }
-
-    fn index_set(&self) -> &Self::Ind {
-        &self.index_layout
-    }
-
-    fn iter(&self) -> Self::Iter<'_> {
-        self.data.as_slice().iter()
-    }
-
-    fn iter_mut(&mut self) -> Self::IterMut<'_> {
-        self.data.as_mut_slice().iter_mut()
-    }
-
-    fn len(&self) -> IndexType {
-        self.index_layout.number_of_global_indices()
-    }
-}
-
-
-impl<T: Scalar> SquareSum for LocalIndexableVector<'_, T> {
-    type T = T;
-    fn square_sum(&self) -> <Self::T as Scalar>::Real {
-        self.iter()
-            .fold(<<Self::T as Scalar>::Real>::zero(), |acc, &elem| {
-                acc + elem.square()
-            })
-    }
-}
-
-impl<T: Scalar> Norm1 for LocalIndexableVector<'_, T> {
-    type T = T;
-    fn norm_1(&self) -> <Self::T as Scalar>::Real {
-        self.iter()
-            .fold(<<Self::T as Scalar>::Real>::zero(), |acc, &elem| {
-                acc + elem.abs()
-            })
-    }
-}
-
-impl<T: Scalar> Norm2 for LocalIndexableVector<'_, T> {
-    type T = T;
-    fn norm_2(&self) -> <Self::T as Scalar>::Real {
-        <<Self::T as Scalar>::Real as Float>::sqrt(self.square_sum())
-    }
-}
-
-impl<T: Scalar> NormInf for LocalIndexableVector<'_, T> {
-    type T = T;
-    fn norm_inf(&self) -> <Self::T as Scalar>::Real {
-        self.iter().fold(
-            <<Self::T as Scalar>::Real as Float>::neg_infinity(),
-            |acc, &elem| <<Self::T as Scalar>::Real as Float>::max(acc, elem.abs()),
-        )
-    }
-} */
