+++ sparse-traits/src/lib.rs
@@ -1,13 +1,12 @@
 //! This is a sandbox to test traits for sparse (and more general) operators.
 
 pub mod index_layout;
-pub mod indexable_vector;
 pub mod operator;
 pub mod spaces;
 pub mod types;
+pub mod linalg;
 
 pub use index_layout::*;
-pub use indexable_vector::*;
 pub use operator::*;
 pub use spaces::*;
 pub use types::*;
