+++ sparse-traits/src/linalg/indexable_vector.rs
@@ -1,4 +1,5 @@
 //! An indexable vector is the standard type for n-dimensional containers
+
 use crate::types::{IndexType, Scalar};
 use crate::IndexLayout;
 
@@ -29,27 +30,4 @@ pub trait IndexableVector {
     fn index_layout(&self) -> &Self::Ind;
 }
 
-pub trait Inner {
-    type T: Scalar;
-    fn inner(&self, other: &Self) -> crate::types::Result<Self::T>;
-}
 
-pub trait SquareSum {
-    type T: Scalar;
-    fn square_sum(&self) -> <Self::T as Scalar>::Real;
-}
-
-pub trait Norm1 {
-    type T: Scalar;
-    fn norm_1(&self) -> <Self::T as Scalar>::Real;
-}
-
-pub trait Norm2 {
-    type T: Scalar;
-    fn norm_2(&self) -> <Self::T as Scalar>::Real;
-}
-
-pub trait NormInf {
-    type T: Scalar;
-    fn norm_inf(&self) -> <Self::T as Scalar>::Real;
-}
