+++ sparse-traits/src/spaces/inner_product_space.rs
@@ -3,5 +3,5 @@ use super::LinearSpace;
 use crate::types::Result;
 
 pub trait InnerProductSpace: LinearSpace {
-    fn inner<'a>(&self, x: ElementView<'a, 'a, Self>, other: ElementView<'a, 'a, Self>) -> Result<Self::F> where Self: 'a;
+    fn inner<'a>(&self, x: &ElementView<'a, 'a, Self>, other: &ElementView<'a, 'a, Self>) -> Result<Self::F> where Self: 'a;
 }
