+++ sparse-traits/src/spaces/normed_space.rs
@@ -4,5 +4,5 @@ use crate::types::Scalar;
 
 pub trait NormedSpace: LinearSpace {
     /// Norm of a vector.
-    fn norm<'a>(&'a self, x: ElementView<'a, 'a, Self>) -> <Self::F as Scalar>::Real;
+    fn norm<'a>(&'a self, x: &ElementView<'a, 'a, Self>) -> <Self::F as Scalar>::Real;
 }
